Create the MiqServer#role_changes method
This will allow us to determine exactly how each role will be
changed during the role sync process.

We use this information to only restart the failover monitor
when the database operations role is being either added or removed.
The failover monitor startup process will then get a chance
to decide whether or not it should be monitoring subscriptions as well
as database.yml

https://bugzilla.redhat.com/show_bug.cgi?id=1391095