Patch ActiveRecord::Migration#migrate to ensure global region schema consistency
This patch does two things.

1. It maintains the migration list in MiqRegion#migrations_ran as each
   migration is completed
2. It halts the global region migration process if the changes from
   the same migration in all the remote regions have not been replicated

This combination will allow us to ensure that the global region will
never attempt to replicate data which would be inconsistent with the
current schema.

This patch uses SQL directly to do the update because the migration
which adds the migrations_ran column was not able to utilize the attribute
for the newly added column.

We were not able to refresh the attributes cache of the MiqRegion
model from within the #migrate method for some reason, but were
able to detect if the column was present and update it while bypassing
the AR caches.

This seemed to be the more sane route given the issues we've had with
model cache clearing in the past.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1601015