Properly create tenant default group for population
When creating a tenant, if there are already groups in the database,
it would assign a random group.
This would manifest by having a seemingly random group assigned to newly
discovered VMs/Hosts. There can then be privilege issues as the admins
possibly don't have access to the group.

This only happens for customers who had a database before tenancy
was introduced in 2015. It incorrectly assigned during the migration
process in 20151021174140_assign_tenant_default_group

The fix: it will now only assign a tenant group.

https://bugzilla.redhat.com/show_bug.cgi?id=1625788

I'm also adding a migration to automatically fix customer's databases
Since the issue was introduced in a migration in the first place.