Leverage submit_job for role based queueing
https://bugzilla.redhat.com/show_bug.cgi?id=1629945

Because the queue put was specifying a queue_name without a role OR
zone, the zone was being set to MiqServer.my_zone.  By leveraging the
higher level method, submit_job, specifying the service of "reporting"
sets the role and because reporting is a regional role, there's no need
to specify the zone, it will default to the "nil" (any) zone.