Add support for magic check in EvmDatabaseOps
Calls the newly created method `MiqFileStorage.magic_number_for` to
pre-fetch the magic from the file that will be downloaded, and send that
to `PostgresAdmin` as a `database_opts` to avoid checking it on it's
end.

Note:  Shortened a comment in here, but since it is part of the git
history, I am fine with it.  Those willing to dig deep can find it.