Add job_timeout parameter for upgrade_cluster
The motivation here is that we need the job to have one timeout, and the
Ansible role to have another - the job will end if the timeout from the
time it was created will pass, while Ansible playbook will only timeout if it is
stack for too long on one phase. So we will want to set the job timeout
to something huge - 1 year, and set a real timeout through the playbook
extra_vars.

This is part of implementing: https://bugzilla.redhat.com/show_bug.cgi?id=1644605