Use Optimist
Trollop is a deprecated gem, and has been renamed/replaced by Optimist.

This will probably also need to happen in conjunction with
`manageiq-providers-vmware`, since that also has a script that uses
this.

That said, this is not intended to be backported, and should just be
used to remove the deprecation warnings the "proper" way.