CustomButtonSet - don't allow saving if buttons specified in button_order don't exist
without this, code like...

```
cbs.set_data[:button_order] = [1,2,3]
cbs.save!
```

would suceed in saving a CustomButtonSet with that button order, even when some of the buttons don't exist.

Since that would mean that children no longer match the button order, we need to throw.

`throw(:abort)` is the only valid way of stopping the chain in Rails 5,
and seems to work find in older rails too.