Fix the metric rollup spec changing the timestamp
The test for updating a metric rollup record keeps the same primary key
was changing the timestamp of the record.  On the first day of the month
this would change which table the record was in, e.g. on Mar 1st the
record would be in metric_rollups_03 and `timestamp - 1.day` would put
it in metric_rollups_02 causing an inheritance check constraint
violation.

Since we never change the timestamp of a record this test is invalid
anyway and the solution is to change some other property which we do
regularly update like a cpu metric value.