Don't start another refresh worker while another is stopping
If a RefreshWorker is asked to stop (usually for exceeding memory) we
don't want to start another right away until the first has stopped.

This can lead to many refresh workers running at the same time since the
next worker that starts typically exceeds memory as well and will be
asked to shutdown also.

In addition to using a huge amount of memory
this can also lead to inventory integrity issues since it is possible
for multiple refresh workers to be running save_ems_inventory at once.