[Rbac::Filterer] Fix Container* belongsto filter
`ContainerGroup`, among other classes, were not being filter based on
it's Entitlements.

This not only includes `ContainerGroup` in the `Rbac::Filterer`'s
`BELONGSTO_FILTER_CLASSES` constant, but also makes sure it is one of the
classes that is filtered via associations, and not as itself and it's
descendants.

Applies the same treatment to all `Container*` classes.