[chargeback_helper] faster .add_metric_rollups_for
This does a lot, so I will try and talk about each change:

To start, it expands the number of iterations through the time range
loop by first mapping the `times` to a variable.  This does mean that
instead of looping only once through the times, it is an "N+1" number of
times (once to generate, and then an iteration for each `resource`), but
this hardly amounts to the slow portions of this method, so it was a
minor concession.

The next change is that the attributes are mostly consistent for each
resource, so by making the `resources.each` the top level loop, we can
cache the values for the record attributes for the entire duration of
each record, and only have to update the `time` when looping through the
`times`.

In addition to the above, the `args` was split out from what as the
`params` to separate the `FactoryBot` factory type and traits from the
record attributes hash.  This further allows the `metric_rollup_params`
to remain constant throughout each `resource` and allows the `.compact`
being called to only have to be applied once.

The rest of the changes revolve around changing the N+1 query down to a
single `INSERT INTO metrics_rollups ( ... ) VALUES ( ... ), ...` call.
This goes outside of `ActiveRecord`'s API to do it, but pretty much
drops the spec times for this file in half and is the bulk for where the
preformance improvements come from:

Before
------

real    1m40.128s    1m44.307s    1m41.907s
user    1m22.867s    1m26.823s    1m24.623s
sys     0m6.500s     0m6.563s     0m6.459s

After
-----

real    0m56.138s    0m58.035s    0m56.721s
user    0m43.822s    0m44.958s    0m44.437s
sys     0m5.603s     0m5.498s     0m5.495s