Disable and unset the subscription slot if the publisher is unreachable
Before this change, the failover monitor would not be able to remove
the subscription when the remote region failed over because it
couldn't remove the replication slot.

To deal with this we need to disable the subscription, set the
replication slot to the special value "NONE" then retry the delete.

We also don't want to do this on every delete because in every other
case we want the remote replication slot to be removed when the
subscription is removed.

More details can be found in the PostgreSQL documentation:
https://www.postgresql.org/docs/10/sql-dropsubscription.html