specify virtual_delegate types to avoid deadlock
deriving the attribute type for delegates required the target
class to be loaded.
This forced a cascade of load_schema calls that end up
introducing a race condition.

This PR explicitly declares the attribute type so the target class
no longer needs to be loaded and the race condition
(and subsequent deadlocks) are avoided.

https://bugzilla.redhat.com/show_bug.cgi?id=1714615
This is based upon:

- https://github.com/ManageIQ/activerecord-virtual_attributes/pull/21
- https://github.com/ManageIQ/manageiq/pull/18798