Add a transaction around the multi-step error-case subscription delete
This is the situation where the remote server is unreachable so
we need to set the replication slot to "NONE" in order to drop the
subscription without error. But we don't want to end up in an
in-between case where the subscription is disabled, but not deleted
so wrap the three commands in a transaction.